refactor(nextjs-apollo-client): rearrange package structure
