+++ packages/nextjs-apollo-client/src/apollo/safeJSON.js
