+++ packages/nextjs-apollo-client/src/apollo/util.ts
